fix(h>0): QR/exp en fp32 explicite — le set_default_dtype(bf16) du transform 8B cassait le QR CPU (geqrf bf16 non implémenté)

+++ aloepri/ffn_obfuscation.py
@@ -42,7 +42,8 @@ def obfuscate_ffn_layer(gate_proj, up_proj, down_proj, seed):
 
     gen = torch.Generator().manual_seed(seed)
     # scaling strictement positif pour rester inversible sans changer de signe
-    scale = torch.exp(torch.randn(intermediate_size, generator=gen) * 0.1)
+    scale = torch.exp(torch.randn(intermediate_size, generator=gen,
+                                    dtype=torch.float32) * 0.1)
 
     gate_proj_obf = gate_proj[perm_index]
     up_proj_obf = up_proj[perm_index] * scale[perm_index].unsqueeze(1)
